requirements.txt
streamlit==1.33.0
groq==0.5.0
python-dotenv==1.0.1

langchain==0.1.20
langchain-community==0.0.38
langchain-huggingface==0.0.3

chromadb==0.4.24
pypdf==4.2.0
sentence-transformers==2.6.1

 